Can't create foreign key for companion_class_schedule on test db

+++ server/db/migrations/20180727174538_companion_class_schedule.js
@@ -4,7 +4,7 @@ exports.up = function (knex, Promise) {
         table.dropUnique(['user_id', 'start_time', 'status'])
 
         table.unique(['user_id', 'start_time', 'status', 'remark'])
-        if (process.env.NODE_ENV !== 'test') {
+        if (process.env.NODE_ENV !== 'test' && process.env.NODE_ENV !== 'qa') {
             table.foreign('user_id').references('users.user_id').onDelete('CASCADE').onUpdate('CASCADE')
         }
     })
